Swap instanbul for nyc & rename some files
