+++ .gitignore
@@ -1,4 +1,7 @@
-node_modules/
-coverage/
+node_modules/**
+build/**
+coverage/**
+.nyc_output/**
+
 .DS_Store
 npm-debug.log*
